Improved character width calculus on Win32s
@roytam1
Again might be interesting to you. Could not test it on a non ANSI localized Win32s but it works on XP using the [Microsoft AppLocal](https://www.softpedia.com/get/System/System-Miscellaneous/Microsoft-AppLocale.shtml) forcing the locale to something like /L0404 (traditional Chinese) or Japanese (L0411) locale...
It is not too slow because the size is only asked once for each Unicode character.