Improve UTF16/32 detection patterns
In an UTF16 file we will always have less than 50% NULs and in an UTF32 file we will always have less than 75% NULs
This help detection and avoids the false positive UTF32 detection on null only files.
Also I directly go to Latin-1 mode if the file contains NULs but is not UTF16/32 because I generally prefer Latin-1 mode to display binary files. Also I want to avoid having a random cp to be selected on binary file because it can be much slower.